A0-4575: Add ABFT performance scorer (#1884)
# Description

Adds a scorer that computes how many rounds which ABFT member was
behind.

## Type of change

- New feature (non-breaking change which adds functionality)

# Checklist:

- I have added tests
- I have created new documentation

+++ pallets/committee-management/src/lib.rs
@@ -274,7 +274,7 @@ pub mod pallet {
         #[pallet::weight((T::BlockWeights::get().max_block, DispatchClass::Operational))]
         pub fn set_finality_ban_config(
             origin: OriginFor<T>,
-            minimal_expected_performance: Option<u32>,
+            minimal_expected_performance: Option<u16>,
             underperformed_session_count_threshold: Option<u32>,
             ban_period: Option<EraIndex>,
         ) -> DispatchResult {
